#6: Fix error when loading project's existing Pull Requests
When running a scan, the scanner engine loads existing Pull Request details, but fails to deserialise the response due to a mismatch between the returned data structure and the target model. Changing the deserialisation method to parse the impacted date field to a Date object using a Date Formatter, and then convert the parsed date to the target model's `long` type resolves this issue.