#19: Add support for SonarQube 7.8 APIs
SonarQube 7.8 contains changes to the APIs required for managing branches, with new methods having been introduced, and the existing `getMergeBranchUUid` method signature in `org.sonar.ce.tasj.projectanalysis.Branch` modified in a non-backwards-compatible way. Due to the incompatible changes, the supported SonarQube API version has been bumped from 7.4 to 7.8 and any methods that only existed to support older SonarQube versions have been removed. The 7.8 release of SonarQube does not currently have a sonar-plugin-api dependency released with it, so the SonarQube 7.7 dependency for the plugin-api component is being used in its place.