Add support for updated `BranchConfigurationLoader` interface
As part of SONAR-12068, the method defined in the `BranchConfigurationLoader` interface was replaced with an alternative method so that implementors are required to only read project-level settings rather than instance-level settings when configuring branches. To support this change, an implementation for this new method has been added to `CommunityBranchConfigurationLoader` whilst the old method definition has been left in-place to allow backwards-compatibility with SonarQube 7.8.