Fix test failure on Java 9 and above
Java 9 modified classloading to use modules, so the default classloader presented by JUnit is not a `URLClassloader` and therefore fails the type check in `ElevatedClassLoaderFactory`. As Sonarqube always creates a classloader for each plugin during the load of that plugin, the exposing of the default App Classloader in the unit test is therefore invalid, so switching to create a custom `URLClassloader` overcomes this. Similarly, Mockito throws an exception when running in Java 9 and above on Windows due to classes eing generated in the wrong module scope. Upgrading to the latest version of Mockito overcomes this.