#38: Add support for SonarQube 8.0
Replaces references to the removed `ScannerUtils` class with Java's `URLEncoder`, and adds implementations of new methods introduced into SonarQube 8.0's interfaces. Since SonarQube 8.0 also changed `ScannerWsClient` from a class to an interface, the `ScannerWsClient` is now invoked through reflection in `ScannerWsClientWrapper` to prevent a `java.lang.IncompatibleClassChangeError` being thrown during invocation.

On top of this, SonarQube Compute Engine now contains a Java Security Manager that blocks access to retrieving `ClassLoader`s, so prevents the work-around used to access the branch API classes required for this plugin to execute. Moving the loading of the impacted classes into a `CoreExtension` and bypassing the class-loading manipulation for these classes overcomes this problem. This will require users of this plugin to install the plugin to both the plugins directory and the core platform's `lib/common` directory (or symlink the files) given plugin classes are only searched for from the plugins directory, and extensions are only loaded from the `common` directory.