Add compile-time support for checking Sonarqube feature deprecation
The existing method of adding comments to methods to indicate what version of Sonarqube the method is supporting is difficult to maintain and prevents external developers from easily working with Sonarqube updates. Adding a new `SonarqubeCompatibility` interface with extensions for each major and minor Sonarqube version that removes any methods, as well as switching existing methods to use these new interfaces, allows for the code-base to consistently compile against different Sonarqube versions, and indicates why any out-dated methods are being retained. This change also corrects tests that incorrectly depended on interfaces or methods to perform a test, where an alternative approach could be used for a more reliable and maintainable result.