Add support for decorating pull requests
Provides the interfaces and models to allow collection of metrics and issues for publishing to pull requests. To allow consistent rendering of information across different version control providers, an `AnalysisDetails` class exposing the collected metrics, and a pre-built message containing a summary of the analysis results. To allows for different providers using different formats of status messages, the status message is produced as a syntax tree, with each Pull Request Provider implementation providing a formatter to convert the tree to a suitable output type. A default Markdown formatter has been included to support this.

The images used for portraying the measures in the decoration have been copied unmodified from the sonarqube/sonarcloud-github-static-resources repository and, as no license is set on this repository or referenced in any of the files, these are presumed to be in the public domain so available for use here.