#62: Don't throw an error if coverage or duplication information is missing
The code previously assumed that Duplication and Coverage information would always be present in the scan results, so always invoked `getValue` on the Quality Gate's `Condition`. However, in the scenario where coverage information isn't available, the `status` of the `Condition` is `NO_VALUE`, and any call to `getValue` results in an `IllegalStateException` being thrown.

Filtering on the condition's status to exclude any condition with `NO_VALUE` results in these fields being treated as `null` so causes the 'No Coverage Information' or 'No Duplication Information' icons to be shown in the analysis summary for these scenarios.