#154: Submit annotations in batches of 50 to Github
Creating a check run on Github limits the number of annotations in the initial report to a maximum of 50, otherwise the request is silently ignored with no error returned in the response. To overcome this, the initial report is limited to 50 annotations, with any remaining annotations submitted in groups of 50 in subsequent update calls.