#130: Allow administrators to define pull request icons source URL
Some ALMs (such as Github) proxy images in Pull Request decoration and therefore need direct access to the images. If the Sonarqube server isn't internet facing, the ALM will not be able to retrieve the images, and they therefore wont be visible on the Pull Request. This change restores the property for allowing Sonarqube administrators to define the URL for retrieving images, with the images defaulting to loading from the Sonarqube if the property is not set/overridden.