#154: Correctly report more than 50 issues to Github
The previous fix for any pull request that had more than 50 issues had allowed an initial check run to be created, and then submitted the next batch of 50 issues to Github but failed to parse the response properly, so threw an error at that point, subsequently only submitting 100 issues to Github for each scan before causing the Compute Engine task to fail. The correct type is now being specified for the parse of Github's response, and unnecessary fields in the expected response object have been removed to allow the response to be parsed properly.