#67: Link back to the Pull Request from Sonarqube
Sonarqube provides the ability to link back to the original Pull Request or Merge Request from the Pull Request overview screen if the Pull Request URL is persisted alongside the branch details. To allow this, decorators are now required to return a `DecorationResult` that contains an optional URL for the Pull Request's overview/summary page. The Gitlab and Github decorators have been updated to provide the new details, however the Bitbucket decorator defaults to returning no URL since the details are not readily available, so Bitbucket Pull Requests will continue to show no link back the the Pull Request from Sonarqube.