Use the correct properties when decorating Bitbucket Pull Requests
As part of the migration so using the new Sonarqube 8.1 APIs, the order of the ALM Slug and ALM Repo are being injected the wrong way round into the Bitbucket URL, so the project is not being found. Switching the properties to the correct order overcomes this issue and allows Pull Request decoration to complete.