#325: Fix NumberFormat call on a non Number
The Gitlab Merge Request decorator was attempting to format the code coverage using a Decimal Format, but not taking into account that some changes do not record any coverage - such as when a non-code change is made - and therefore represents the coverage as a `null` value. This change alters the handling of the coverage values in the Gitlab decorator to treat `null` coverage values as `0`.