#334: Add ability to collect Bitbucket HTTP debug logs
When Bitbucket Pull Request decoration fails due to an error from Bitbucket, diagnosing the cause of the failure can be challenging due to the lack of logs. This change therefore adds an interceptor to the HTTP client used in Bitbucket decoration which logs HTTP request headers and body content to the Sonarqube logs at debug level.