Prevent BitBucket unit tests initiating real HTTP calls
The BitBucket cloud decorator was attempting to perform an OAuth negotiation during construction in unit tests, so was dependent on the availability of BitBucket for the tests to pass. The `BitbucketClientFactory` has been altered to take an OkHttp Client Builder Supplier as an argument, with the tests injecting a mock client builder so no Http calls are performed during the BitBucket decorator tests.