#272: Ensure New Code Period settings are persisted
The New Code Period webservices contain logic that specifically sets the target branch UUID to the main branch of the project if the Community version of Sonarqube is being used, thereby overriding the branch name that's been saved in the new code period's value.

The existing Java Agent functionality has been extended to alter the constructors of the relevant New Code services so they override the bundled Edition Provider that specifies the current platform edition as being Community edition, and instead injects a provider that overrides the current distribution as Developer edition, thereby causing the webservice logic to skip applying defaults for branch values.