Fix bad usages of Mockito classes
Some tests currently use Mockito to create mock classloaders, which do not then call `super` properly so cause classes to be defined with no module and cause subsequent JVM crashes when these classes are used later in the testsuite on newer Mockito versions.

An internal Mockito class has also been incorrectly referenced in a test in place of a Sonarqube class with a similar name, and a spy setup incorrectly around a mock, both of which fail in the latest Mockito version. These have both been updated to work correctly regardless by referencing the correct Sonarqube class and using the mock directly, rather than attempting to spy an already mocked instance.