Update to latest Gradle version
The Gradle Wrapper has been reconfigured to use Gradle 7.1.1, and the wrapper files regenerated. Deprecated configuration for XML Jacoco reporting has been updated to the new standard for Gradle 7, and the removed scopes (compile, runtime) have been migrated to their equivalent (implementation, runtimeOnly), as well as a general cleanup of dependency definition to make the formatting consistent. To ensure the download of the Sonarqube Zip archive continues to work, the Ivy repository definition has been explicitly configured to use the target artefact as the metadata, rather than defaulting to searching for an Ivy XML metadata file that isn't present as part of the target distribution.