#442: Pass preview flag on Azure Devops status requests
Azure DevOps on-premise responds with an error if the `-preview` is not appended to the version number during the request to submit the pipeline status during Pull Request decoration. As the use of a preview version for this call does not seem to adversely affect invocation of this endpoint on Azure DevOps Cloud, the `-preview` flag has been unconditionally added to the status update, although it has not been added to any other invocations since they do not encounter the same error.