#569: Add support for release 9.4 of Sonarqube
The latest release of Sonarqube switches from using Pico Container for
dependency injection to using Spring. The method used for detecting the
constructor to be used for injecting in a multi-constructor class varies
between the 2 frameworks, with Pico only looking for public constructors
but Spring looking for any visible constructor, and then excluding any
constructors that do not have the `@Autowired` annotation on a muti-
constructor class. This change adds the relevant annotation onto the two
classes impacted by this change in frameworks, thereby allowing for the
Compute Engine to have the appropriate components generated for
completing decoration of Pull Requests.