Replace null values in Bitbucket analysis report
Bitbucket throws an error when a report contains a percentage element
with a null value. The code coverage and duplication metrics are now
nullable in the report data passed around the plugin, bit the Bitbucket
decorator does not check or replace nulls from these fields.

This change alters the report publishing to check for nulls in either of
the coverage or duplication fields and uses Zero in their place.