#508: Support Pagination over AppInstallation
When more than a single page worth of app installation is present in the
Github API, the plugin does not currently fetch beyond the initial page,
and therefore doesn't find the access tokens for the required
installation. By loading all pages of installations before searching for
the current application we ensure that we don't ignore any relevant
installation.