#674: Add support for Sonarqube 9.7
Sonarqube has removed the use of the same external component key and
differing database IDs for different branches on the same projects and
now use different IDs in all cases. The pull request web service
endpoints have also been removed from community edition. To allow the
plugin to work with the new version of Sonarqube, the component key
generation for branches has been modified to save branch DTOs whenever a
new branch is created, and to remove the conditions around re-using the
same branch if the target branch details matched an existing branch. The
Pull Request endpoint actions have been copied from the old community
edition sources, and tidied up to use a cleaner abstraction model. As
the front-end only shows branch features if the implementation of
`BranchFeatureExtension` returns the name 'branch-support', the
`CommunityBranchFeatureExtension` has been altered to follow this
requirement, and an additional `MonoRepoFeature` has been implemented to
allow the mono-repo switches to be shown against the front-end.

Includes the migration of any altered unit tests to JUnit 5.