#704: Add support for Sonarqube 9.8
The retrieval of the list of pull requests has been removed from the
scanner engine, as well as the issue metrics on the Pull Request web
service. The scanner has therefore been altered to implement the updated
APIs and drop the classes that implemented the removed interfaces, and
the Pull Request web service has had the gathering and population of the
issue metrics removed.