Clean-up retrieval of Github authentication tokens
The current implementation of the RestApplicationAuthenticationProvider
contains an instance-level Object Mapper but doesn't use it instead
constructing a new Object Mapper on each invocation and eagerly loads
all pages of App Installations even though the first page may contain a
token with access to the required repository. The implementation has
therefore been altered to use the constructed object mapper for all
operations and to work through the full contents of a page
of AppInstallations before moving on to the next page of installations.