Clean-up unit tests for CommunityBranchPlugin class
The tests for CommunityBranchPlugin were using argument captors against
repeated calls to the varargs `addExtensions` method, but were only
verifying some of the calls, and were not taking into account the types
passed into the methods. Version 5 of Mockito has a slight change in how
it handles varargs capturing, so will not correctly match the varargs
invocations where the same matcher has been used to try and match a
subset of arguments, or where the method is only using one argument for
the varargs section. By switching to verifying all the values directly
on each method call the tests now provide meaningful coverage as well as
allowing for a clean upgrade to Mockito 5.