Fix reference license in project root
The project contains LGPL headings in all the source files and is
intended to be licensed as an LGPL project, but contains a GPL license
file in the project root. As contributors have been committing with LGPL
headers this switch is being made without any wider changes across the
project.